requirement.txt
python-magic
python-magic-bin==0.4.14
pathlib